Convert Tokeniser to a single-line Reader format.
A Tokeniser no longer takes []byte on each pass, nor does it
return [][]string.  Instead, it takes an io.Reader, and spins
on it until it either errors or hits a line.  It then returns
this one line, holding a buffer of things it's read from the
Reader to use in subsequent tokenisations.

This is still experimental, and will break your stuff.  The
connector has been hopefully repaired to work with the new
Tokeniser, but caveat emptor.