Simplify tokeniser internals.
This changes the tokeniser semantics a bit--we're no longer doing
Unicode replacement.  This means test U2 no longer passes and is
disabled.  However, I think we've talked about this(?) and this
is actually a good thing--playd AFAIK doesn't do it, and it
mangles otherwise perfectly-good-but-not-quite-UTF8 messages in
a way that breaks messages moving through bifrost-go-using middle
servers in a stack.

The spec needs editing to remove U2, then.

This will perform slower than before, but shouldn't be a problem.