Remove MessageWords in favour of strings.
This simplifies the message implementation dramatically, allows
non-standard messages to be routed by bifrost-go, and will allow
further simplifications down the road.

Most of the Rs* and Rq* constants still exist, but now point
directly to their former `.String()` equivalents.

If your code relied on BadWord, RsUnknown, RqUnknown, etc., or
MessageWord being a distinct thing, it is now broken.  If it only
used the other Rq* and Rs* constants, it may be fine.