Replace with 3rd-gen Bifrost protocol work
This work comes from the `baps3d` repository, which will soon change to
point here instead.  (As I understand it, _that_ work was based on the
work from _this_ repository.)

All of the various resource tree work has been deleted, as the current
direction of the bifrost protocol doesn't use it.

Apologies to everyone who put in work on the old protocol - your
sacrifice, I hope, was not in vain.